차트 업데이트
pH문제 해결 (원인: 그래프의 데이터 기준선은 무조건 LAxis)

+++ 0201_메인서버_및_APP/app/navmain/app/src/main/java/com/example/shinj/navmain/Temperature/TemperaturePresenterImpl.java
@@ -1,5 +1,6 @@
 package com.example.shinj.navmain.Temperature;
 
+import android.util.Log;
 import android.widget.EditText;
 
 import io.socket.client.Socket;
@@ -9,14 +10,14 @@ public class TemperaturePresenterImpl implements TemperaturePresenter {
 
     @Override
     public void saveTemperPH(Socket socket, EditText minTemper, EditText maxTemper, EditText minPH, EditText maxPH) {
-
-        int minTemperValue = Integer.parseInt(minTemper.getText().toString());
-        int maxTemperValue = Integer.parseInt(maxTemper.getText().toString());
+        double minTemperValue = Double.parseDouble(minTemper.getText().toString());
+        double maxTemperValue = Double.parseDouble(maxTemper.getText().toString());
         double minPHValue = Double.parseDouble(minPH.getText().toString());
         double maxPHValue = Double.parseDouble(maxPH.getText().toString());
         socket.emit("insertTemper", minTemperValue, maxTemperValue);
         socket.emit("insertPH", minPHValue, maxPHValue);
     }
+
 
     @Override
     public void initTemperPH(Socket socket) {
